+++ tango-dataview/src/game/bn5/save.rs
@@ -476,7 +476,7 @@ impl<'a> save::AutoBattleDataViewMut<'a> for AutoBattleDataViewMut<'a> {
             return false;
         }
         self.save.buf[0x7340 + id * std::mem::size_of::<u16>()..][..std::mem::size_of::<u16>()]
-            .copy_from_slice(bytemuck::bytes_of(&count));
+            .copy_from_slice(bytemuck::bytes_of(&(count as u16)));
         true
     }
 
@@ -485,7 +485,7 @@ impl<'a> save::AutoBattleDataViewMut<'a> for AutoBattleDataViewMut<'a> {
             return false;
         }
         self.save.buf[0x2340 + id * std::mem::size_of::<u16>()..][..std::mem::size_of::<u16>()]
-            .copy_from_slice(bytemuck::bytes_of(&count));
+            .copy_from_slice(bytemuck::bytes_of(&(count as u16)));
         true
     }
